Splitting up hearing schedule actions
