+++ client/app/hearings/constants.js
@@ -6,7 +6,6 @@ export const ACTIONS = {
   RECEIVE_HEARING: 'RECEIVE_HEARING',
   RECEIVE_SAVED_HEARING: 'RECEIVE_SAVED_HEARING',
   RESET_SAVE_SUCCESSFUL: 'RESET_SAVE_SUCCESSFUL',
-  CANCEL_HEARING_UPDATE: 'CANCEL_HEARING_UPDATE',
   RECEIVE_HEARING_DAY_OPTIONS: 'RECEIVE_HEARING_DAY_OPTIONS',
   RECEIVE_UPCOMING_HEARING_DAYS: 'RECEIVE_UPCOMING_HEARING_DAYS',
   REGIONAL_OFFICE_CHANGE: 'REGIONAL_OFFICE_CHANGE',
