+++ client/app/hearings/containers/AssignHearingsContainer.jsx
@@ -12,7 +12,7 @@ import {
   onReceiveUpcomingHearingDays,
   onSelectedHearingDayChange,
   onReceiveAppealsReadyForHearing
-} from '../actions/actions';
+} from '../actions/hearingScheduleActions';
 import { onRegionalOfficeChange } from '../../components/common/actions';
 import LoadingDataDisplay from '../../components/LoadingDataDisplay';
 import { COLORS, LOGO_COLORS } from '../../constants/AppConstants';
